Fix CSS customization docs
I was trying to customize my Jekyll blog's CSS and found that the path for the main scss file (for minima 2.5.0) is assets/main.scss, and that's what you have to recreate in your own Jekyll repo to get customizations to work. It is not /assets/css/style.scss (creating this does nothing).